refactor : [005-SELLER_SIGN_UP_IN] 프로젝트 폴더 구조 변경
service 코드 customer 과 seller 로 구분
